Networking: Default gateway support
Discover the first valid IPv4 enabled interface within the
container namespace, connect it to the integration bridge
and pass the IP address, subnet and the default gateway associated
with this interface into the container using kernel command
line parameters.

If the container has multiple IPv4 enabled interfaces only the
first interface will be discovered and added to the container.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>