build: Don't use realpath
realpath is not always available, for instance it isn't in the debian
image used by travis. Close inspection of the build log showed:

  ./configure: line 14573: realpath: command not found

Instead, and to not introduce another dependency, use cd + pwd. Note
that the tests/functional/bundle directory doesn't exist in a pristine
checkout, hence the stderr redirection.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>