Scripts: Add helper script to allow options to be passed to runtime.
Docker 1.12 allows an alternative OCI runtime to be specified, but does
not allow options to be passed to that alternate runtime. A solution to
this problem is to configure Docker to use /usr/bin/cc-oci-runtime.sh
and have that script invoke the actual runtime binary with the users
preferred options (such as "--global-log" and maybe "--debug").

Signed-off-by: James Hunt <james.o.hunt@intel.com>