oci: Force enum oci_namespace to be a signed int
Compiling with clang, we had:

src/namespace.c:109:14: warning: implicit conversion changes signedness: 'int'
      to 'enum oci_namespace' [-Wsign-conversion]
                        return p->num;

The problem is cc_oci_map tries to map any enum to a string using an
int, but the signedness of enums can change depending on the values they
store. enum cc_oci_status is a signed int while enum cc_oci_status is an
unsigned int

By forcing INVALID to -1, both of those enums have the same signedness,
matching the one in cc_oci_map.

Looking the users of those enums, it seems we are always comparing
ns->type with OCI_NS_INVALID, so that should be ok.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>