docs: containerd seems to do the "start" and "delete" calls
A close look at containerd(-shim) seems to indicate it's only
responsible for the "create" command, leaving the other ones to
containerd.

shim:
  containerd-shim/process.go	->	create, exec

containerd:
  runtime/process.go		->	start,
  runtime/container.go		->	delete, state, ps, events, ...

To be thorough if anyone wants to double check the above:
containerd-shim/process.go also includes a "delete" call, but it's only
used if "create" fails, so not the normal mode of operation.

I also modified the life time of the shim accordingly, it exits right
after reaping the qemu process, leaving the delete for containerd.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>