Networking: Move network discovery and configuration to child
Move the network configuration and the discovery process to
the child. The parent runs the hooks and on completion of the hooks
notifies the child of the hook execution status. The child will launch
the VM if the hooks were executed successfully, or will exit.

This reduces the complexity and the latency in the network setup.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>