Version: Ensure "--version" and "version" display a commit id.
git-archive(1), as used by github.com, allows files to be modified
"on-the-fly" as the archive file is generated.

By configuring git-archive to write the commit id to file
"commit_id.fmt", the build can be convinced (using the mandatory
"clean-local" rule) to generate file "commit_id" from "commit_id.fmt".
The value of "commit_id" is then passed via the compiler to the runtime
to allow the latter to display a commit id when run with either
"--version" or "version".

Signed-off-by: James Hunt <james.o.hunt@intel.com>