Process: Support running hooks with no arguments.
The cc_run_hook() function sets the G_SPAWN_FILE_AND_ARGV_ZERO flag
when spawning a hook using g_spawn_async_with_pipes().

The precise semantics of this flag are not clear from the glib
documentation. What the doc should state is that if the flag is set it expects:

  argv[0] = full path of the command to run.
  argv[1] = name of the command to run.
  argv[2] = next argument to pass to argv[0], or NULL.

However, in the case where cc_run_hook() was used to run a hook with no
arguments, it was setting...

  argv[0] = full path of the command to run.
  argv[1] = NULL.

This resulted in an error message and the hook failed to run. The
solution is to simply remove the G_SPAWN_FILE_AND_ARGV_ZERO flag since
then g_spawn_async_with_pipes() expects:

  argv[0] = full path of the command to run.
  argv[1] = next argument to pass to argv[0], or NULL.

Note that strictly the argv vector was still incorrect when a hook with
1+ arguments was executed, but glib tolerates this scenario.

Signed-off-by: James Hunt <james.o.hunt@intel.com>