Tests: Split tests so each calls run() only once.
The "run" function in the latest release of bats(1) has a subtle bug
whereby it is not valid to call run() twice in the same @test function.
Doing so will result in bizarre behaviour due to IFS being modified.

One solution is to surround each call to run with code to save/restore
"$IFS", but that looks ugly. A neater solution is to ensure that each
@test function only calls run once.

Signed-off-by: James Hunt <james.o.hunt@intel.com>