Config: Remove bug report email address.
The address specified in "configure.ac" is not that useful given that
there is now a mailing list that could be used instead, as specified in
CONTRIBUTING.md. However, that document already explains how we want
bug reports to be handled (github issues) so a bug report address is
redundant.

Signed-off-by: James Hunt <james.o.hunt@intel.com>