cmdline: Clarify where you can place global and command options
The Usage string inferred that you could place any options any side
of the [command], but what it meant to indiate was that global options
come before the command and command specific options come after the
command. Change the message to make that clearer.

Ideally we would also modifiy the 'Options:' paragraph section in the
help to say 'Global Options:', but that is buried in the source of the
glib g_option source, with no way to override.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>