Config: Remove "--with-tests-rootfs-path=".
There is already a "--with-tests-bundle-path=" option so also having
"--with-tests-rootfs-path=" option is overkill since the rootfs is
canonically found in a "rootfs/" sub-directory of the bundle directory.

The "--with-tests-rootfs-path=" option was being used to generate
"config-minimal-cc-oci.json", so by removing the option from
configure.ac it is necessary to handle that generation in Makefile.am.

This is accomplished by generalising the rule to create
"cc-oci-runtime.sh" such that the rule now works for all files listed in
GENERATED_FILES.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>