Config: Generate bundle directory by default.
The configure script allows the specification of a bundle directory
("--with-tests-bundle-path=") for the tests to make use of.

However, the default should be to generate the bundle automatically.
This is now done using "make-bundle-dir.sh". If automatic bundle
generation is not desirable, specify "--disable-auto-bundle-creation".

The bundle will be generated by default below "tests/functional/bundle/"
but that default can be modified by specifying
"--with-auto-bundle-creation-path=".

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>