Config: Expand path for automatic bundle-generation.
The automatic bundle generation feature (commit 7dd2c5a) contained a
bug where the default automatic bundle path was relative rather than
absolute.

Now, both the default and the path specified by
"--with-auto-bundle-creation-path=" is made absolute.

Note that symbolic links are not expanded, for symmetry with
other paths passed to configure (such as "--with-cc-image=" and
"--with-cc-kernel=" where symlink expansion is expressely undesirable).

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>