start: Correct how we determine if the runtime is being run standalone
The correct way of determining if we are being run by conatinerd during
start phase is to check if no tty is attached to stdin. We should'nt
be checking for config.terminal value as this refers to the stdin terminal
for the create phase.
Revert earlier changes introduced with this bug. As a side note, the terminal
value will be stored in the state file in the future to account for
the code that determines the config.terminal value programmatically.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>