travis: Always run the tests from an exploded make dist tarball
We run the travis script from an exploded `make dist` tarball to ensure
`make dist` has the necessary files to compile and run the tests.

Note: we don't use `make distcheck` here as we can't run everything we
      want for discheck in travis.

This commit introduces chronic, a useful little utility that will only
print stdout/err if the command fails.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>