Logging: Update global log first.
If global logging is enabled, write to that log before the standard log.

This ordering change guards against the scenario where if the runtime
misbehaves somehow and the container is then deleted, the container and all it
state (including its log) will be deleted.

But if the runtime is still in fact running, it can no longer log to the
standard log (since the log directory will no longer exist), but that write
failure means the global log will not be written to either.

By changing the write order, the global log will always be available,
regardless of the state of the container.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>