cc-shim: Change stdin handling to account for raw mode tty.
Poll returns immediately after a character is entered when the
terminal is in raw mode and not when enter/return is hit.
Buffer the characters until newline is received.

The way docker sets up the tty, we do not get a carriage return,
but have kept this in case this changes.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>