tests/mock: Fix race between Add(1) and Wait() on Workgroup
Turns out, we can't call Wait() on a Workgroup with counter 0 and then
try to Add(1). In retrospect it makes sense, there is something
unsettling to have Wait() called with counter 0 (it shouldn't wait)
while calling Add(1) (instructing Wait() to wait).

Reference: https://golang.org/pkg/sync/#WaitGroup.Wait

"Note that calls with a positive delta that occur when the counter is
zero must happen before a Wait. Calls with a negative delta, or calls
with a positive delta that start when the counter is greater than zero,
may happen at any time. Typically this means the calls to Add should
execute before the statement creating the goroutine or other event to be
waited for."

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>