proxy: Add a "containerId" argument to Bye
Archana suggested we add a containerId parameter to Bye so we don't have
to issue an attach payload to associate the connection beforehand. That
saves a round-trip with the proxy.

Because it's still early times and we don't have any user, let alone any
stable user, yet, I broke the API: the old behaviour using Bye with no
parameter when the client is already attached to a VM will not work
anymore.

Suggested-by: Archana Shinde <archana.m.shinde@intel.com>
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>