Tests: Unbreak bats error handling.
Previously, if certain bats helper functions failed, they would log a
message and "exit 1". This was incorrect since "exit 1" caused the
script to exit, meaning the bats framework was unaware that an error
had occured. The result was that no error output would be produced for
failing tests.

Replacing the "exit 1" with a call to false(1) results in the correct
behaviour since bats notices the failing assertion triggered by the
false result and displays the error as expected, before failing the
test.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>