console: Fix prompt when bash is run interactively.
When we run bash in interactive mode under docker, the prompt is
not shown. We can see the prompt for other shells.

A quick test using this:
https://www.gnu.org/software/bash/manual/html_node/Is-this-Shell-Interactive_003f.html

shows that bash is not running in interactive mode causing the
systemwide startup files not being read and settings not
being applied resulting in PS1 variable to be empty.

The shell is not run in interactive mode because stderr is not
pointing to a terminal even though stdout and stdin are attached to
a tty by hypertstart. The reason being, hyperstart does not attach
the terminal to tty when the stderr sequence number is not set to 0.

Hence, allocate just one stream for the tty case, and explicitly set the
error sequence number to 0 with the "newcontainer" command.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>