create: exit from  infinite while loop
File descriptos 0,1,2 are close on exec.
This allow shim use stdio fd to print and read.

The fds are duplicated until its value is more
than 2.
If close the old file descriptor, next time
dup is called will get the lowest fd , and this
could be the previous closed, causing that
runtime never exit from while loop.

After this loop exec is done, and previous duplicated
fd are close on exec.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>