cor: start command waits for cc-shim ends
Since cc-oci-runtime must be able to run containers
standalone, start command is in charge of determinate
if shim is running or not also start command must
update state file once shim has finished.

This patch opens a file and locks it after fork and before
exec cc-shim when cc-shim ends this file will be unlocked
allowing to start command catch that event and update the state file,
this is another way to detect when a process ends.

Signed-off-by: Julio Montes <julio.montes@intel.com>