oci: Start pause container
In theory, we should not have to start the pause container as
the hyperstart pod is equivalent to the pause container k8s
and CRI-O are using.

In practice, hyperstart kills its pod when the last remaining
container is done. So we can't launch 2 containers in a row if
we don't tell hyperstart to launch the pause container as part
of the pod.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>