gitignore: add missing cscope files
When using cscope command the default output is the file "cscope.out",
and when using the parameter -f on cscope, the generated index will add
an .in and .po to the -f argument. i.e. the followig command

cscope -b -f cscope.out

will generate

cscope.out
cscope.out.in
cscope.out.po

Signed-off-by: Miguel Bernal Marin <miguel.bernal.marin@linux.intel.com>