tests: fix cc_oci_setup_shim test
This patch fixes a random issue for cc_oci-setup_shim unit
tests.

The issue was caused because a new function parameter
was added but the function definition was not updated
in the unit test. As a result the new parameter value
(int shim_flock_fd) got random values because was not properly
initialized and the test failed when the value was negative.

The function definition is updated according to the function
signature.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>