ci: Make the gofmt test output the guilty source files
Because gofmt doesn't return a useful exit status code when finding a
problem but does output the list of guilty source files, we count the
number of outpout lines and error out if we more than 0.

The problem is that we lose the information of which file needs fixing.
Using tee allows us to display that information to the developer.

    Doing go static checks on packages: ./proxy ./proxy/api
    Running misspell...
    Running go vet...
    Running gofmt...
    /home/damien/go/src/github.com/01org/cc-oci-runtime/proxy/fdleak_test.go
    Makefile:5230: recipe for target 'check-go' failed

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>