vendor: Revendor virtcontainers to 0.2.3 to fix a potential panic
Let's vendor virtcontainers 0.2.3, it's a good version if we want to
minimize the changes while grabbing a fix for a panic that happens if
the shim disappears underneath us.

It's debatable if we should jump straight away to 0.3.0. That version
would also need us to re-vender the runv package. It will have to
happen, eventually, but we can control when.

An benefit of going slowly on the updates is that we can revert a later
vendoring to 0.3.0 if we find out it introduces regressions.

Fixes: https://github.com/01org/cc-oci-runtime/issues/463

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>