proxy: Make Hello and Attach return the current api version
Let's version the proxy API so we can document API additions as we
formalize things a bit better.

Note that version 1 of the protocol is due to be frozen with the 2.1
release so it should be considered in flux until then.

After that, we should be able to evolve the protocol is a backward
compatible way.

Fixes: https://github.com/01org/cc-oci-runtime/issues/574
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>