proxy: Support build the proxy without linker flags
At the moment we rely on a linker flag to provide the path to the socket
to create when the proxy is started from the command line (ie. not using
systemd).

We can provide a fallback path in line with what we do in the Makefile
to support the common way of building go programs by just invoking

$ go build

While at it, we can reduce the scope of the socketPath variable a bit.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>