proxy: Don't quit the hyper->client IO goroutine when a shim dies
That's a bug we have had for the longest time. There is one such
goroutine per VM and we where quitting it whenever we had an error
writing data to a shim.

However, we may have a number of shim per-VM, one for the initial
container and one for each exec. We can't just quite the whole goroutine
if the write to one of the shim is failing.

This bug was discovered because we know signal "VM lost" when this
goroutine exits, which was causing havoc.

Fixes: https://github.com/01org/cc-oci-runtime/issues/638
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>