tests: Tighten the docker info check
Since:

  commit c8b933adb193d611fc0998368f866d8028efe99a
  Author: Julio Montes <julio.montes@intel.com>
  Date:   Thu Jan 26 18:16:44 2017 -0600

      tests: fix docker functional tests

We don't run docker rm in setup(), which is a good thing and avoid hiding bugs.

However it's possible the system we are running on still has some
(stopped) containers not yet deleted (I had some of those on my system
and the test was failing).

Instead of checking the number of known containers, check the number of
running container. Add a check on the status of the created container as
well to make sure that one is indeed running.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>