tests: Swarm - get replicas only from active containers
The swarm tests are verifying the interfaces and IP addresses
from the containers listed with 'docker ps -a -q', but this fails
if there is a container that is not running (which maybe was created
before running the tests).

This commit changes the command to 'docker ps -q' so that we only
get the list of containers that are running.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>