ovs-dpdk: Add support for ovs-dpdk
Runtime determines that ovs-dpdk is desired if the IP for the interface
matches the address appended to a /var/run/openvswitch/ovd_ file.
If this port exists and matches, we should take necessary actions
to setup the guest for making use of OVS and DPDK.

In the even that we are using OVS-DPDK, we will need to provide
a mounted hugetablefs for the VM to use.  As an example:
    -object memory-backend-file,id=dpdk-mem,size=1024M,mem-path=/dev/hugepages,share=on,prealloc=on \
    -numa node,memdev=dpdk-mem

On the networking side, an example:
    -chardev socket,id=char1,path=/run/openvswitch/vhost-user1 \
    -netdev type=vhost-user,id=mynet1,chardev=char1,vhostforce \

Fixups:
-don't use -memprealloc parameter. memprealloc will zero out all
the memory, including nvdimm.  Bummer.  Instead, just use
prealloc=on for the numa node

Signed-off-by: Eric Ernst <eric.ernst@intel.com>