Build: Fix -race pre-test to test cc-proxy
The code introduced to test if -race is supported on commit
19c45031ddf04240bbbb9748cbb35e100c22ef3f was not working.
Ubuntu OBS builds fails at that point after detecting an error.

With this simple fix, the build will try to test cc-proxy with -race
enabled and if it fails it will try another time without -race.

Fixes #670

Tested on OBS builds.

Signed-off-by: Geronimo Orozco <geronimo.orozco@intel.com>