shim: Always close console tty if it was opened.
cc_oci_setup_shim() opens the console device and sets the standards
streams to use the associated fd. If the original open fails, the
function returns failure meaning that the tty fd should always be
closed after a sucessful open. Previously, this was only happening
when no further operation failed.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>