docs: Add DAX appendix to architecture documentation
We have DAX information on the clearlinux.org website that could do
with being in this architecture doc.
The DAX info is a specific technical detail, so feels better to be
in an appendix rather than in the main body of the text.
Also add a simple DAX diagram, and update the master diagram ODP
file as well.

Also, whilst extending the TOC, indent it with 4 spaces instead of
2, as it helps render sub-bullets properly with other tools (such
as pandoc)

Signed-off-by: Graham Whaley <graham.whaley@intel.com>