Docs: README: Add GO requirements
As we now have golang code in the tree, you must have the repo/code
in your GOPATH, or it won't build. Add that info to the Building
section.
To go along with that, also add go as a package requirement, and
note the minimum version. And sort that package list whilst there.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>