networking: Save and flush iptable rules in the net namespace
Iptable rules are set by Docker in Swarm mode in the network
namespace. As the container within the VM is not aware of these,
we need to save these and propagate them to the VM.
As the rules get applied to the packets crossing the tap bridge,
we need to flush them in the network namespace.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>