docs: Install: Fedora: Fix broken markdown hotlink
A hotlink was using '[] ()', where the inter-bracket space breaks
markdown rendering of the hotlink on github at least. Fix it (remove
the space).
And, subsequently, also fix up some line length issues in the source
of that paragraph and some wordology at the same time.

Fixes #857 (issue is referenced in that ticket)

Signed-off-by: Graham Whaley <graham.whaley@intel.com>