tests: common: Add new method for saving results
Add a new function (save_results) for the metrics tests to migrate
to use, which then calls out to a script to do the any actual
processing and saving of the results.

This is the first step on formalising, documenting and unifying
the results reporting method across all the metrics tests.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>