tests: Add function to check when a swarm service is ready
This patcha adds a function to be used in tests related with swarm.

The function check_swarm_replicas: Will wait until all the replicas
from a docker swarm service is ready. It also uses a timeout
if after the timeout is finished the function will fail.

The function uses :

```
docker service ps SERVICE
```
instead of:

```
docker ps --filter status=running --filter ancestor=IMAGE
```

This is needed because docker 17 does not filter containers that
are part of a swarm service.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>