configure: Add warning when kernel or image is not found
This patch adds a warnings at configuration time, it will
show a warning message when either clear containers kernel
or image are not found.

It also check if a kernel or image was installed by clear containers
packages. If they are installed it will use them as default paths. This
is useful for people that want to use cc-oci-runtime compiled from
source but had a previous installation from packages.

Partially Fixes: #769

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>