sriov: Add network namespace SRIO-V support
Overview of SRIO-v enabling code in cc-oci-runtime:

 1. Before discovering network, create a HASH of mac to BDF for each
network device (each device listed in /sys/class/net).

 2. During network discovery, check to see if mac address is within the
hash for each interface.  If it is, then we assume VF device pass-through.
If it is, note the BDF, device/vendor ID and the original driver. In the event
the network device is VF based, don't do any further network setup. After
network setup, the mac-to-BDF hash is freed.

3. During QEMU command line parameter preperation, for each VF interface,
unbind it from the original host host driver and bind it do vfio-pci driver,
and add appropriate command line specific to using VFIO interface.

4. When creating state file, add entries into the JSON for
the device, and specifically for the network device.

5. During teardown: when shutdown command is received, the stop handler
will bind the device driver back to the host and unbind the given VF
device. We also remove the device ID from the VFIO-PCI driver so new virtual
functions aren't automatically bound to that driver.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>