mounts: Bind mount container rootfs to workload directory
Bind-mount the container rootfs to workload dir for a standalone
container. The container rootfs will now live under
{runtime_dir}/cid/workload/cid/rootfs.

Refactor cc_handle_mounts() to mount to the workload directory
in case we are mounting a rootfs and to mount to the workload
directory after generating a unique name for a non-rootfs mount.
The actual mounting to the correct location under the rootfs
will be taken care by hyperstart. We do need to pass this info
in fsmap structure.
Add host_path to cc_oci_mount for this purpose, it will hold
the unique name generated above. The current mnt.mnt_dir field
holds the actual filename. Store these to the state file so that
they can be passed with the newcontainer command to hyperstart.

Since we now use a hyperstart shared location for regular
container as well, collapse the mounting logic in cc_handle_mounts
for pods and regular containers to just check for volumes.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>