storage: Add function to check the storage for container rootfs
If a block device is found, we store the device name of the block
device found. We also store the file system type and index.
We are currently using an incrementing counter for the index.
When we add support for CRIO with hot plugging drives, so that
containers with block drives can be added after the VM is started,
we should store the global index in a file and retrieve
it from there.

The index is needed to derive the drive name that we pass to qemu.
We need to pass the drive name and file system type to hyperstart.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>