storage: Pass drive name to hyperstart with the newcontainer command
For hyperstart to use the drive passed on qemu command line,
we pass the drive name and the file system type of the block
device. The drive name needs to be passed as image field
in the newcontainer json.
The drive name is deduced based on the index at which the drive is
passed to qemu.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>